+++ ui/prochandlers.go
@@ -14,6 +14,7 @@ import (
 	"github.com/gcla/gowid"
 	"github.com/gcla/gowid/widgets/table"
 	"github.com/gcla/termshark/v2"
+	"github.com/gcla/termshark/v2/configs/profiles"
 	"github.com/gcla/termshark/v2/pcap"
 	log "github.com/sirupsen/logrus"
 )
@@ -120,7 +121,7 @@ func (t updatePacketViews) OnError(code pcap.HandlerCode, app gowid.IApp, err er
 		fmt.Fprintf(os.Stderr, "%v\n", err)
 		RequestQuit()
 	} else {
-		if !termshark.ConfBool("main.suppress-tshark-errors", false) {
+		if !profiles.ConfBool("main.suppress-tshark-errors", false) {
 			var errstr string
 			if kverr, ok := err.(gowid.KeyValueError); ok {
 				errstr = fmt.Sprintf("%v\n\n", kverr.Cause())
@@ -153,7 +154,7 @@ func (t SimpleErrors) OnError(code pcap.HandlerCode, app gowid.IApp, err error) 
 	log.Error(err)
 	// Hack to avoid picking up errors at other parts of the load
 	// cycle. There should be specific handlers for specific errors.
-	if !termshark.ConfBool("main.suppress-tshark-errors", false) {
+	if !profiles.ConfBool("main.suppress-tshark-errors", false) {
 		app.Run(gowid.RunFunction(func(app gowid.IApp) {
 			OpenError(fmt.Sprintf("%v", err), app)
 		}))
@@ -372,7 +373,7 @@ func (s SetStructWidgets) OnError(code pcap.HandlerCode, app gowid.IApp, err err
 	// Hack to avoid picking up errors at other parts of the load
 	// cycle. There should be specific handlers for specific errors.
 	if s.Ld.PdmlLoader.IsLoading() {
-		if !termshark.ConfBool("main.suppress-tshark-errors", false) {
+		if !profiles.ConfBool("main.suppress-tshark-errors", false) {
 			app.Run(gowid.RunFunction(func(app gowid.IApp) {
 				OpenLongError(fmt.Sprintf("%v", err), app)
 			}))